+++ examples/grid.rs
@@ -5,6 +5,7 @@ use pushrod::render::callbacks::widget_id_for_name;
 use pushrod::render::engine::Engine;
 use pushrod::render::widget::Widget;
 use pushrod::render::widget_config::{CONFIG_BORDER_WIDTH, CONFIG_COLOR_BORDER, CONFIG_COLOR_TEXT};
+use pushrod::render::{make_points, make_size};
 use pushrod::widgets::grid_widget::GridWidget;
 use pushrod::widgets::slider_widget::SliderOrientation::SliderHorizontal;
 use pushrod::widgets::slider_widget::SliderWidget;
@@ -38,14 +39,21 @@ pub fn main() {
         .build()
         .unwrap();
     let mut engine = Engine::new(400, 340, 60);
-    let mut grid1 = GridWidget::new(20, 20, 360, 280, 10);
+    let mut grid1 = GridWidget::new(make_points(20, 20), make_size(360, 280), 10);
 
     grid1
         .get_config()
         .set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     grid1.get_config().set_numeric(CONFIG_BORDER_WIDTH, 1);
 
-    let mut slider1 = SliderWidget::new(20, 310, 320, 20, 1, 20, 10, SliderHorizontal);
+    let mut slider1 = SliderWidget::new(
+        make_points(20, 310),
+        make_size(320, 20),
+        1,
+        20,
+        10,
+        SliderHorizontal,
+    );
 
     slider1.on_value_changed(|_slider, _widgets, _layouts, pos| {
         let text1_id = widget_id_for_name(_widgets, String::from("text1"));
@@ -61,10 +69,8 @@ pub fn main() {
         16,
         TextJustify::Left,
         String::from("10"),
-        360,
-        310,
-        40,
-        20,
+        make_points(360, 310),
+        make_size(40, 20),
     );
 
     text_widget1