+++ examples/layout.rs
@@ -11,6 +11,7 @@ use pushrod::render::widget_cache::WidgetContainer;
 use pushrod::render::widget_config::{
     PaddingConstraint, CONFIG_BORDER_WIDTH, CONFIG_COLOR_BORDER, CONFIG_COLOR_TEXT,
 };
+use pushrod::render::{make_points, make_points_origin, make_size};
 use pushrod::widgets::push_button_widget::PushButtonWidget;
 use pushrod::widgets::text_widget::{TextJustify, TextWidget};
 use sdl2::pixels::Color;
@@ -42,28 +43,28 @@ pub fn main() {
     let mut layout = HorizontalLayout::new(20, 20, 360, 80, PaddingConstraint::new(0, 0, 0, 0, 1));
     let mut layout2 =
         VerticalLayout::new(250, 120, 130, 160, PaddingConstraint::new(0, 0, 0, 0, 1));
-    let mut widget1 = BaseWidget::new(0, 0, 0, 0);
+    let mut widget1 = BaseWidget::new(make_points_origin(), make_size(0, 0));
 
     widget1
         .get_config()
         .set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     widget1.get_config().set_numeric(CONFIG_BORDER_WIDTH, 2);
 
-    let mut widget2 = BaseWidget::new(0, 0, 0, 0);
+    let mut widget2 = BaseWidget::new(make_points_origin(), make_size(0, 0));
 
     widget2
         .get_config()
         .set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     widget2.get_config().set_numeric(CONFIG_BORDER_WIDTH, 2);
 
-    let mut widget3 = BaseWidget::new(0, 0, 0, 0);
+    let mut widget3 = BaseWidget::new(make_points_origin(), make_size(0, 0));
 
     widget3
         .get_config()
         .set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     widget3.get_config().set_numeric(CONFIG_BORDER_WIDTH, 2);
 
-    let mut widget4 = BaseWidget::new(0, 0, 0, 0);
+    let mut widget4 = BaseWidget::new(make_points_origin(), make_size(0, 0));
 
     widget4
         .get_config()
@@ -88,10 +89,8 @@ pub fn main() {
         16,
         TextJustify::Right,
         String::from("Spacing:"),
-        20,
-        116,
-        70,
-        22,
+        make_points(20, 116),
+        make_size(70, 22),
     );
 
     text_widget1
@@ -104,17 +103,20 @@ pub fn main() {
         16,
         TextJustify::Left,
         String::from("1"),
-        100,
-        116,
-        40,
-        22,
+        make_points(100, 116),
+        make_size(40, 22),
     );
 
     text_widget2
         .get_config()
         .set_color(CONFIG_COLOR_TEXT, Color::RGB(0, 0, 255));
 
-    let mut button1 = PushButtonWidget::new(130, 112, 50, 30, String::from("<"), 20);
+    let mut button1 = PushButtonWidget::new(
+        make_points(130, 112),
+        make_size(50, 30),
+        String::from("<"),
+        20,
+    );
 
     button1.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button1.set_numeric(CONFIG_BORDER_WIDTH, 2);
@@ -146,7 +148,12 @@ pub fn main() {
         cast!(_widgets, text_widget2_id, TextWidget).set_text(format!("{}", spacing));
     });
 
-    let mut button2 = PushButtonWidget::new(180, 112, 50, 30, String::from(">"), 20);
+    let mut button2 = PushButtonWidget::new(
+        make_points(180, 112),
+        make_size(50, 30),
+        String::from(">"),
+        20,
+    );
 
     button2.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button2.set_numeric(CONFIG_BORDER_WIDTH, 2);
@@ -184,10 +191,8 @@ pub fn main() {
         16,
         TextJustify::Right,
         String::from("Top:"),
-        20,
-        146,
-        70,
-        22,
+        make_points(20, 146),
+        make_size(70, 22),
     );
 
     text_widget3
@@ -200,17 +205,20 @@ pub fn main() {
         16,
         TextJustify::Left,
         String::from("0"),
-        100,
-        146,
-        40,
-        22,
+        make_points(100, 146),
+        make_size(40, 22),
     );
 
     text_widget4
         .get_config()
         .set_color(CONFIG_COLOR_TEXT, Color::RGB(0, 0, 255));
 
-    let mut button3 = PushButtonWidget::new(130, 142, 50, 30, String::from("<"), 20);
+    let mut button3 = PushButtonWidget::new(
+        make_points(130, 142),
+        make_size(50, 30),
+        String::from("<"),
+        20,
+    );
 
     button3.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button3.set_numeric(CONFIG_BORDER_WIDTH, 2);
@@ -242,7 +250,12 @@ pub fn main() {
         cast!(_widgets, text_widget4_id, TextWidget).set_text(format!("{}", top));
     });
 
-    let mut button4 = PushButtonWidget::new(180, 142, 50, 30, String::from(">"), 20);
+    let mut button4 = PushButtonWidget::new(
+        make_points(180, 142),
+        make_size(50, 30),
+        String::from(">"),
+        20,
+    );
 
     button4.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button4.set_numeric(CONFIG_BORDER_WIDTH, 2);
@@ -280,10 +293,8 @@ pub fn main() {
         16,
         TextJustify::Right,
         String::from("Bottom:"),
-        20,
-        176,
-        70,
-        22,
+        make_points(20, 176),
+        make_size(70, 22),
     );
 
     text_widget5
@@ -296,17 +307,20 @@ pub fn main() {
         16,
         TextJustify::Left,
         String::from("0"),
-        100,
-        176,
-        40,
-        22,
+        make_points(100, 176),
+        make_size(40, 22),
     );
 
     text_widget6
         .get_config()
         .set_color(CONFIG_COLOR_TEXT, Color::RGB(0, 0, 255));
 
-    let mut button5 = PushButtonWidget::new(130, 172, 50, 30, String::from("<"), 20);
+    let mut button5 = PushButtonWidget::new(
+        make_points(130, 172),
+        make_size(50, 30),
+        String::from("<"),
+        20,
+    );
 
     button5.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button5.set_numeric(CONFIG_BORDER_WIDTH, 2);
@@ -338,7 +352,12 @@ pub fn main() {
         cast!(_widgets, text_widget6_id, TextWidget).set_text(format!("{}", bottom));
     });
 
-    let mut button6 = PushButtonWidget::new(180, 172, 50, 30, String::from(">"), 20);
+    let mut button6 = PushButtonWidget::new(
+        make_points(180, 172),
+        make_size(50, 30),
+        String::from(">"),
+        20,
+    );
 
     button6.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button6.set_numeric(CONFIG_BORDER_WIDTH, 2);
@@ -376,10 +395,8 @@ pub fn main() {
         16,
         TextJustify::Right,
         String::from("Left:"),
-        20,
-        206,
-        70,
-        22,
+        make_points(20, 206),
+        make_size(70, 22),
     );
 
     text_widget7
@@ -392,17 +409,20 @@ pub fn main() {
         16,
         TextJustify::Left,
         String::from("0"),
-        100,
-        206,
-        40,
-        22,
+        make_points(100, 206),
+        make_size(40, 22),
     );
 
     text_widget8
         .get_config()
         .set_color(CONFIG_COLOR_TEXT, Color::RGB(0, 0, 255));
 
-    let mut button7 = PushButtonWidget::new(130, 202, 50, 30, String::from("<"), 20);
+    let mut button7 = PushButtonWidget::new(
+        make_points(130, 202),
+        make_size(50, 30),
+        String::from("<"),
+        20,
+    );
 
     button7.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button7.set_numeric(CONFIG_BORDER_WIDTH, 2);
@@ -434,7 +454,12 @@ pub fn main() {
         cast!(_widgets, text_widget8_id, TextWidget).set_text(format!("{}", left));
     });
 
-    let mut button8 = PushButtonWidget::new(180, 202, 50, 30, String::from(">"), 20);
+    let mut button8 = PushButtonWidget::new(
+        make_points(180, 202),
+        make_size(50, 30),
+        String::from(">"),
+        20,
+    );
 
     button8.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button8.set_numeric(CONFIG_BORDER_WIDTH, 2);
@@ -472,10 +497,8 @@ pub fn main() {
         16,
         TextJustify::Right,
         String::from("Right:"),
-        20,
-        236,
-        70,
-        22,
+        make_points(20, 236),
+        make_size(70, 22),
     );
 
     text_widget9
@@ -488,17 +511,20 @@ pub fn main() {
         16,
         TextJustify::Left,
         String::from("0"),
-        100,
-        236,
-        40,
-        22,
+        make_points(100, 236),
+        make_size(40, 22),
     );
 
     text_widget10
         .get_config()
         .set_color(CONFIG_COLOR_TEXT, Color::RGB(0, 0, 255));
 
-    let mut button9 = PushButtonWidget::new(130, 232, 50, 30, String::from("<"), 20);
+    let mut button9 = PushButtonWidget::new(
+        make_points(130, 232),
+        make_size(50, 30),
+        String::from("<"),
+        20,
+    );
 
     button9.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button9.set_numeric(CONFIG_BORDER_WIDTH, 2);
@@ -530,7 +556,12 @@ pub fn main() {
         cast!(_widgets, text_widget10_id, TextWidget).set_text(format!("{}", right));
     });
 
-    let mut button10 = PushButtonWidget::new(180, 232, 50, 30, String::from(">"), 20);
+    let mut button10 = PushButtonWidget::new(
+        make_points(180, 232),
+        make_size(50, 30),
+        String::from(">"),
+        20,
+    );
 
     button10.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button10.set_numeric(CONFIG_BORDER_WIDTH, 2);