+++ src/render/widget_cache.rs
@@ -172,7 +172,7 @@ impl WidgetCache {
     /// When state is set to `true`, this indicates that a mouse button down was detected.  When set
     /// to `false`, it indicates that the mouse button was released.  When setting the button state
     /// to `widget_id == -1`, the button click message will be sent to _all_ `Widget`s, so use
-    /// `widget_id == - 1` with care.
+    /// `widget_id == -1` with care.
     pub fn button_clicked(
         &mut self,
         widget_id: i32,