+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Pushrod Change Log
 
+## 0.4.21
+
+- Fixed typo in mod.rs (#263)
+- Removed logging from push button and image button.
+- Converted `xywh` coordinates to `Point` and `Size` (#266)
+- Add callback when Event::Quit is detected (#210)
+- Added default of `true` to exit the application if no quit callback is set. (#210)
+- Added `exit` application to demo exiting with a prompt. (#210)
+
 ## 0.4.20
 
 - Sample applications cleanup